requirements.txt
# Core dependencies
opencv-python>=4.5.5.64
torch>=1.10.0
torchvision>=0.11.1
numpy>=1.21.0
Pillow>=9.0.0

# Data processing dependencies
glob2>=0.7
tqdm>=4.62.3
scikit-learn>=1.0.2
pathlib>=1.0.1
zipfile36>=0.1.3
elementpath>=2.5.0

# YOLOv5 dependencies
matplotlib>=3.5.1
PyYAML>=6.0
requests>=2.27.1
scipy>=1.7.3
seaborn>=0.11.2
pandas>=1.3.5

# Optional - for performance improvement
# Uncomment if using CUDA
# cupy-cuda11x>=10.4.0
